+++ client/src/pages/UserInfo.jsx
@@ -11,35 +11,19 @@ import Button from 'react-bootstrap/Button';
 
 
 function UserInfo() {
-    return ( <
-            div >
-            <
-            JumboSection / >
-            <
-            Row >
-            <
-            Col sm = "12"
-            md = "4" >
-            <
-            UserForm / >
-            <
-            /Col> <
-            Col sm = "12"
-            md = "8" >
-            <
-            ShopperStores / >
-            <
-            /Col> {
-            /* <Col sm="12" md="8">
-                                <GroupStoreView />
-                            </Col> */
-        } <
-        /Row>
+    return (<div>
+        <JumboSection />
+        <Row>
+            <Col sm="12" md="4" >
+                <UserForm />
+            </Col>
+            <Col sm="12" md="8" >
+                <ShopperStores />
+            </Col>
+        </Row>
 
-    <
-    /div>
-
-)
+    </div>
+    )
 }
 
 export default UserInfo;
